more refactoring with mono
