+++ src/ir_passes/mono/mod.rs
@@ -3,4 +3,7 @@ mod monomorphize;
 mod monosig;
 mod utils;
 
+pub(super) use monosig::MonoSig;
+pub(super) use utils::{Base, BaseComp, IntoBase, Underlying, UnderlyingComp};
+
 pub use monomorphize::Monomorphize;
