+++ src/ir_passes/mono/utils/mod.rs
@@ -1,0 +1,7 @@
+mod base;
+mod comp;
+mod underlying;
+
+pub use base::{Base, IntoBase};
+pub use comp::{BaseComp, UnderlyingComp};
+pub use underlying::Underlying;
